backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
celery==5.3.4
pandas==2.1.4
numpy==1.26.4
yfinance==0.2.28
requests==2.31.0
aiohttp==3.9.1
TA-Lib==0.6.7
scikit-learn==1.3.2
textblob==0.17.1
cloud-sql-python-connector==1.4.3
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.1
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.38.1
google-cloud-redis==2.15.1
vertexai>=1.38.0
newsapi-python==0.2.7
praw==7.7.1
vaderSentiment==3.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-benchmark==4.0.0
pytest-html==4.1.1
pytest-timeout==2.2.0
coverage==7.3.2
factory-boy==3.3.0
freezegun==1.2.2
httpx==0.25.2
pydantic-settings==2.1.0
reportlab==4.4.3
weasyprint==66.0
jinja2==3.1.6